requirements.txt
requests==2.31.0
python-dotenv==1.0.0
eth-utils==2.3.1
eth-abi==5.0.1 